Set the default monospace font for fontconfig.
When Firefox is set up to not allow webpages to choose their own fonts instead of the configured fonts, it uses the configured monospace font everywhere, except in `pre` tags, where it queries fontconfig for the monospace font to use. The response from fontconfig may not be the same as the configured monospace font. This file tells fontconfig that when a monospace font is requested, it should respond with my preferred font.