MSYS2 multi-byte rendering bug: used simpler symbols (#40)
The previous workaround was writing the command timer report to standard output, so it became a part of `PS1`. (This somehow causes multi-byte characters to get rendered correctly.) But if the command contains backticks, the text within them is interpreted as a shell command when `PS1` is parsed. Hence, the report is again written to standard error, but simpler symbols are used on MSYS2.