Handle colouring of full-path completion words
The completion words for `chsh -s` are full paths to shell programs (e.g. `/usr/bin/zsh`). Suppose I typed `chsh -s /usr/` and pressed Tab, `/usr/` in the completion words would not be coloured grey, because the completion colouring rule I had configured tried to colour the base name of `/usr/` (which is the empty string) in grey. For normal file name completion, colouring only the base name is the correct behaviour (because only base names are displayed as completion words), but for this special `chsh` use case, allow colouring the entire typed word if it is found in any completion word as a prefix.