Enabled shell integration, but disabled configurable features
I don't want shell integration. I like the shell and the terminal to remain separate entities, so I configure the shell I use to behave the way I like. However, Ghostty 1.0.0-dev+0000000 requires shell integration to open a new tab in the same directory as the previously-focused tab. Hence, I am forced to enable it. Fortunately, it allows to disable all shell integration features (where 'all' excludes the feature to open a new tab in the same directory), so I did just that.