Tried something; did not work
The problem with the `--dir-diff` option is that Git just gives us two directories without any information on renames. The diff tool thus sees the rename as one added file and one deleted file. To solve this, I thought I could remove that option and parse each diff that Git gives us, and keep appending the results to an HTML file. This isn't possible because the diff tool is invoked once for each diff pair, and does not know when the current diffing task has ended and another one has started. Git creates different temporary directories for each diff pair, so we can't use the names of the temporary directories, either.