Specify encoding for `FileInput` (#105)
It uses CP-1252 on Windows, which chokes on the multi-byte characters I have used with abandon in my projects. A funny bug results: files containing such characters are marked as binary by `diff.py`, even if they are text files. This happens because the exception raised is the same one I catch to detect when a binary file is opened.

Also, `locale` is now imported in the Python REPL.